chore: move util function a bit more
