+++ react-components/src/data-providers/core-dm-provider/CoreDm3dDataProvider.ts
@@ -27,7 +27,7 @@ import { filterNodesByMappedTo3d } from './filterNodesByMappedTo3d';
 import { getCadModelsForInstance } from './getCadModelsForInstance';
 import { getCadConnectionsForRevisions } from './getCadConnectionsForRevisions';
 import { zip } from 'lodash';
-import { restrictToDmsId } from '../utils/restrictToDmsId';
+import { restrictToDmsId } from '../../utilities/restrictToDmsId';
 
 const MAX_PARALLEL_QUERIES = 2;
