+++ react-components/src/data-providers/core-dm-provider/getDMSRevision.ts
@@ -4,7 +4,7 @@
 import { type QueryRequest } from '@cognite/sdk';
 import { type DmsUniqueIdentifier, type FdmSDK, type NodeItem } from '../FdmSDK';
 import { type COGNITE_CAD_REVISION_SOURCE, type CogniteCADRevisionProperties } from './dataModels';
-import { restrictToDmsId } from '../utils/restrictToDmsId';
+import { restrictToDmsId } from '../../utilities/restrictToDmsId';
 import { revisionQuery } from './revisionQuery';
 
 export async function getDMSRevision(
